pyproject.toml
[project]
name = "lambdalayerproject"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aws-cdk-lib>=2.170.0",
    "boto3>=1.42.73",
    "constructs>=10.0.0",
]
